feat: Add a new output s3_bucket_versioning_id
This new output will allow having an implicit dependency between the versioning and the resources that need it outside the module (an S3 bucket object for example).

The caller will then reference this output instead of `s3_bucket_id` (depending on if it uses the `expected_bucket_owner` or not).